Add Corrections
Corrections on MakeTransparentGameObject Script :
The transparence is now applied on gameObjects which have children. All
the children become transparent when the parent is detected between the
player and the camera.

Correction on the MechanismBase script and all mechanisms : The Base
script now contains the "Destroy(this)". We do not need to destroy the
script inherited from MechanismBase anymore, this will be done
automatically when the Activate method is called.